requirements.txt
attrs==22.2.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mistune==2.0.5
mysqlclient==2.1.1
packaging==23.0
protobuf==4.22.1
PyJWT==2.6.0
pyrsistent==0.19.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3
